Enable Prettier w/ ESLint and add setup in README
Prettier is my love. The one that you never knew you needed till you lose it. I can’t code without it! When you `cmd + s` to save… and magic… so nice.

In the future, for those users in other editors we can add in a commit hook to auto-format. These can get annoying at times though. Might be worth experimenting with the team.