Add Routing and setup initial routes
We’re bringing in react-router to handle our SPA routing. RR is a very popular routing library and great for SPAs for it’s simplicity, support, and size.

I started with 3 routes… “/login”, “/lists”, and “/lists/:listId”. We will need to build in authentication to handle private routes and add in some redirects to keep the User from hitting auth/unauth’d pages when they are unauth’d/auth’d.