Create Login page
The Login page takes in the User’s email and password, authenticates their credentials via the API, and if successful, saves their auth token as a cookie. The User is then redirected into the App.

I had the option of saving the auth token as a cookie or in local storage. After some very quick research, I found a recommendation to save as a cookie for increased security. I’m not an expert in authentication, so others might have better opinions. Open to feedback.

This commit also adds the “REACT_APP_APP_URL” environment variable. This sets the API URL to be used on the client. To set environment variables, create a .env file at the root of your project (instructions in README).