Extract LoginForm into Feature
We can extract smaller parts of a Page into their own components, which we call features. These features can be re-used across many Pages and breaks larger problems down into smaller solutions. This makes development, testing, and maintainance a little easier.

It’s not required to break a page down into entirely features, nor is it always the best option. In this case, I wanted to show an example of a feature… but it’s not really providing much benefit at this time.