Add Logout
We use the “/logout” route to log the User out. This allows any page, feature, component to direct the User to that route and they will be logged out. Easy!

Our current logout logic doesn’t invalid the token. So it is still available for use if someone were to copy it. It might be a good idea to let the server know that the User is logging out so it can invalidate the token on its side.