Add check for generic error
I handled the 404 and 401 status codes, but the server may respond with others. If we miss the status code check, we alert the User with a generic error.

According to the docs, “400 - when invalid parameters are passed”. I don’t think this should be possible unless we messed up our code or the backend updated it’s parameters.