Create and retrieve Lists
We’ve implemented most of the “/lists” Page. On mount, we fetch the lists from the API and save them in our global state. We can also create new Lists and save them in our global state. They persist between page loads.

We haven’t implemented the tasks yet, so the completion data is hard-coded in the UI right now.