Add ability to complete/incomplete a Task
Toggling a task is now possible. I did notice that the server wasn’t that quick to respond when completing a task. It caused the checkbox to feel a little laggy… so I decided to visually show the task completed before the server responded to our request. I did some light QA to ensure that we didn’t get out of sync.